refactor(service): the assembly leaves cli/, which a public entry may not reach
