+++ test/cli-serve.test.ts
@@ -3,7 +3,8 @@ import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { describe, expect, it, vi } from "vitest";
 import type { Agent } from "../src/agent.ts";
-import { mountAgentcore, mountSessionControl, reportServing, routesFor } from "../src/cli/serve.ts";
+import { mountAgentcore, reportServing } from "../src/cli/serve.ts";
+import { mountSessionControl, routesFor } from "../src/service.ts";
 import { router } from "../src/channels/serve.ts";
 import { text } from "../src/channels/respond.ts";
 import { INVOKE_EXAMPLE_BODY } from "../src/channels/http.ts";
